Logic to implement B2G1 changed in offerAppliedOnMilk method

+++ src/main/java/Item.java
@@ -21,7 +21,7 @@ public class Item {
         return quantity;
     }
 
-    public void setQuantity(double quantity) {
+    public void setQuantity(int quantity) {
         this.quantity = quantity;
     }
